rename dynamic_op to dynamic_ops
