+++ mmrazor/models/architectures/__init__.py
@@ -1,7 +1,7 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 from .backbones import *  # noqa: F401,F403
 from .connectors import *  # noqa: F401,F403
-from .dynamic_op import *  # noqa: F401,F403
+from .dynamic_ops import *  # noqa: F401,F403
 from .generators import *  # noqa: F401,F403
 from .heads import *  # noqa: F401,F403
 from .ops import *  # noqa: F401,F403
