+++ mmrazor/models/mutators/channel_mutator/slimmable_channel_mutator.py
@@ -7,7 +7,7 @@ import torch.nn as nn
 from torch.nn import Module
 from torch.nn.modules.batchnorm import _BatchNorm
 
-from mmrazor.models.architectures.dynamic_op import DynamicBatchNorm
+from mmrazor.models.architectures.dynamic_ops import DynamicBatchNorm
 from mmrazor.models.mutables import SlimmableMutableChannel
 from mmrazor.registry import MODELS
 from ...task_modules import PathList
