+++ mmrazor/models/task_modules/delivery/__init__.py
@@ -1,5 +1,5 @@
 # Copyright (c) OpenMMLab. All rights reserved.
-from .deliver_manager import DistillDeliveryManager
+from .delivery_manager import DistillDeliveryManager
 from .function_outputs_delivery import FunctionOutputsDelivery
 from .method_outputs_delivery import MethodOutputsDelivery
