+++ mmrazor/models/task_modules/delivery/delivery_manager.py
@@ -3,7 +3,7 @@ import copy
 from typing import Dict, Optional
 
 from mmrazor.registry import TASK_UTILS
-from .distill_deliver import DistillDelivery
+from .distill_delivery import DistillDelivery
 
 SUPPORT_DELIVERIES = ['FunctionOutputs', 'MethodOutputs']
