+++ tests/test_models/test_architectures/test_dynamic_op/test_bricks/test_dynamic_conv.py
@@ -8,9 +8,9 @@ import pytest
 import torch
 from torch import nn
 
-from mmrazor.models.architectures.dynamic_op.bricks import (BigNasConv2d,
-                                                            DynamicConv2d,
-                                                            OFAConv2d)
+from mmrazor.models.architectures.dynamic_ops.bricks import (BigNasConv2d,
+                                                             DynamicConv2d,
+                                                             OFAConv2d)
 from mmrazor.models.mutables import OneShotMutableChannel, OneShotMutableValue
 from mmrazor.structures.subnet import export_fix_subnet, load_fix_subnet
 from ..utils import fix_dynamic_op
@@ -233,7 +233,7 @@ def test_kernel_dynamic_conv2d(dynamic_class: Type[nn.Conv2d],
 @pytest.mark.parametrize('dynamic_class', [OFAConv2d, BigNasConv2d])
 def test_mutable_kernel_dynamic_conv2d_grad(
         dynamic_class: Type[nn.Conv2d]) -> None:
-    from mmrazor.models.architectures.dynamic_op.bricks import \
+    from mmrazor.models.architectures.dynamic_ops.bricks import \
         dynamic_conv_mixins
 
     kernel_size_list = [3, 5, 7]
