+++ tests/test_models/test_architectures/test_dynamic_op/test_bricks/test_dynamic_norm.py
@@ -6,10 +6,8 @@ import pytest
 import torch
 from torch import nn
 
-from mmrazor.models.architectures.dynamic_op.bricks import (DynamicBatchNorm1d,
-                                                            DynamicBatchNorm2d,
-                                                            DynamicBatchNorm3d,
-                                                            DynamicMixin)
+from mmrazor.models.architectures.dynamic_ops.bricks import (
+    DynamicBatchNorm1d, DynamicBatchNorm2d, DynamicBatchNorm3d, DynamicMixin)
 from mmrazor.models.mutables import OneShotMutableChannel
 from mmrazor.structures.subnet import export_fix_subnet, load_fix_subnet
 from ..utils import fix_dynamic_op
